requirements.txt
asgiref==3.4.1
backoff==1.10.0
boto3==1.18.9
botocore==1.21.9
click==8.0.1
Deprecated==1.2.12
Flask==2.0.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
opentelemetry-api==1.4.1
opentelemetry-distro==0.23b2
opentelemetry-exporter-otlp==1.4.1
opentelemetry-exporter-otlp-proto-grpc==1.4.1
opentelemetry-instrumentation==0.23b2
opentelemetry-instrumentation-botocore==0.23b2
opentelemetry-instrumentation-dbapi==0.23b2
opentelemetry-instrumentation-flask==0.23b2
opentelemetry-instrumentation-grpc==0.23b2
opentelemetry-instrumentation-logging==0.23b2
opentelemetry-instrumentation-sqlite3==0.23b2
opentelemetry-instrumentation-urllib==0.23b2
opentelemetry-instrumentation-wsgi==0.23b2
opentelemetry-proto==1.4.1
opentelemetry-sdk==1.4.1
opentelemetry-sdk-extension-aws==0.23b2
opentelemetry-semantic-conventions==0.23b2
opentelemetry-util-http==0.23b2
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
